Changes/Comments on Requirment Document
Fixed wording within the Requirement Document.

Also added one extra requirement to **User Interface Requirements for Ingredients:**
"Ability to add Ingredients"

Comments on The Requirement Document:

The User should also be able to input the quantity of the ingredients he has to narrow down the search. Such as if a dish requires 3 eggs but the user only has 2 eggs, it should not be listed or if it is listed should be put as a warning or caution

Should the user be able to create their own type of category??? such as favorite foods and etc?

Should we add a calorie count for ingredients??(for health users)